+++ mobile/screens/__tests__/LoginScreen.test.js
