+++ mobile/screens/__tests__/SignUpScreen.test.js
