+++ mobile/screens/components/__mocks__/SingUp.js
