+++ mobile/screens/tab_navigator/car_defense/CarDefenseApp.js
