+++ mobile/screens/tab_navigator/indica_ai/IndicaAiApp.js
