+++ mobile/screens/tab_navigator/settings/SettingScreen.js
