+++ mobile/screens/tab_navigator/vendas/TabHandlerVendasApp.js
