+++ mobile/screens/tab_navigator/vendas/components/BuyerOrderCard.js
