+++ mobile/screens/tab_navigator/vendas/components/OfferDialog.js
