+++ mobile/screens/tab_navigator/vendas/components/OrderCard.js
