+++ mobile/screens/tab_navigator/vendas/components/ProductCard.js
