+++ mobile/screens/tab_navigator/vendas/components/ProductImage.js
