+++ mobile/screens/tab_navigator/vendas/components/styles.js
