+++ mobile/screens/tab_navigator/vendas/screens/MyProductsScreen.js
