+++ mobile/screens/tab_navigator/vendas/screens/OfferScreen.js
