+++ mobile/screens/tab_navigator/vendas/screens/my_products/ErrorDialog.js
