+++ mobile/screens/tab_navigator/vendas/screens/offers/OfferDetails.js
