+++ mobile/screens/tab_navigator/vendas/screens/offers/Offers.js
